QTi/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
rclone-python==0.1.0
python-binance==1.0.19
ccxt==4.1.87
ta==0.10.2
python-socketio==5.11.1
maturin==1.4.0
pyo3==0.20.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2
joblib==1.3.2
scikit-learn==1.3.2
scipy==1.11.4
redis==5.0.1
paramiko==3.4.0 